worker/requirements.txt
celery
redis
python-dotenv
requests
sqlalchemy
psycopg2-binary
pydantic
watchdog
boto3
instructor
anthropic